backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.22.0
sqlmodel>=0.0.14
alembic>=1.11.1
passlib[bcrypt]>=1.7.4
python-jose[cryptography]>=3.3.0
python-multipart>=0.0.6
psycopg2-binary>=2.9.6
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
